requirements.txt
flask>=2.3.0
langdetect>=1.0.9
yake>=0.4.8
pdfplumber>=0.10.0
newspaper3k>=0.2.8
sumy>=0.11.0
nltk>=3.8.0
deep-translator>=1.11.0
lxml[html_clean]>=4.9.0
